refactor, move all SOAP client classes into one package, GP-7220
