+++ src/main/java/org/genepattern/gpunit/soap/ModuleRunnerSoap.java
@@ -1,4 +1,4 @@
-package org.genepattern.gpunit.exec.soap;
+package org.genepattern.gpunit.soap;
 
 import static org.junit.Assert.fail;
